Rename the test
