+++ tests/wasm/optional_primitives.js
