+++ tests/wasm/optional_primitives.rs
@@ -1,7 +1,7 @@
 use wasm_bindgen_test::*;
 use wasm_bindgen::prelude::*;
 
-#[wasm_bindgen(module = "tests/wasm/optional_numbers.js", version = "*")]
+#[wasm_bindgen(module = "tests/wasm/optional_primitives.js", version = "*")]
 extern {
     fn i32_js_identity(a: Option<i32>) -> Option<i32>;
     fn u32_js_identity(a: Option<u32>) -> Option<u32>;
